Add Product Performance report endpoint and refactor tests
Introduced a new Product Performance report feature, including backend query logic, MediatR handler, and a new API endpoint for retrieving product sales performance by date range. Added supporting models and response types. Refactored and reorganized transaction, merchant, operator, and contract endpoint tests into dedicated files for clarity. No breaking changes to existing endpoints; all previous functionality remains intact.
