+++ EstateReportingAPI.Models/ProductPerformanceResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace EstateReportingAPI.Models;
+
+public class ProductPerformanceResponse
+{
+    public List<ProductPerformanceDetail> ProductDetails { get; set; }
+    public ProductPerformanceSummary Summary { get; set; }
+}
